terraform: fill in more flat interfaces
