+++ terraform/graph_config_node_resource.go
@@ -260,6 +260,13 @@ func (n *GraphNodeConfigResourceFlat) DependentOn() []string {
 		prefix)
 }
 
+func (n *GraphNodeConfigResourceFlat) ProvidedBy() []string {
+	prefix := modulePrefixStr(n.PathValue)
+	return modulePrefixList(
+		n.GraphNodeConfigResource.ProvidedBy(),
+		prefix)
+}
+
 // graphNodeResourceDestroy represents the logical destruction of a
 // resource. This node doesn't mean it will be destroyed for sure, but
 // instead that if a destroy were to happen, it must happen at this point.
