+++ terraform/transform_root.go
@@ -34,3 +34,7 @@ type graphNodeRoot struct{}
 func (n graphNodeRoot) Name() string {
 	return "root"
 }
+
+func (n graphNodeRoot) Flatten(p []string) (dag.Vertex, error) {
+	return n, nil
+}
